+++ .gitignore
@@ -1,6 +1,9 @@
 # Created by https://www.toptal.com/developers/gitignore/api/pycharm+all,visualstudiocode,python
 # Edit at https://www.toptal.com/developers/gitignore?templates=pycharm+all,visualstudiocode,python
 
+# temp files
+*.tmp
+
 # File-based project format
 *.iws
 
@@ -183,4 +186,4 @@ Thumbs.db
 # End of https://www.toptal.com/developers/gitignore/api/pycharm+all,visualstudiocode,python
 
 # timberborn map files
-.timber
+*.timber
