+++ .pre-commit-config.yaml
@@ -1,0 +1,21 @@
+# See https://pre-commit.com for more information
+# See https://pre-commit.com/hooks.html for more hooks
+repos:
+-   repo: https://github.com/pre-commit/pre-commit-hooks
+    rev: v4.3.0
+    hooks:
+    -   id: trailing-whitespace
+    -   id: end-of-file-fixer
+    -   id: check-added-large-files
+        args: ['--maxkb=4096']
+-   repo: https://gitlab.com/pycqa/flake8
+    rev: '5.0.4'  # pick a git hash / tag to point to
+    hooks:
+    -   id: flake8
+        args: ['--max-line-length=128', '--extend-ignore=E127']
+-   repo: https://github.com/pycqa/isort
+    rev: '5.10.1'
+    hooks:
+      - id: isort
+        name: isort (python)
+        args: ['--line-length=128']
